requirements.txt
asgiref==3.2.5
Babel==2.8.0
beautifulsoup4==4.9.1
braintree==4.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
console==0.9901
cryptography==3.0
dj-database-url==0.5.0
Django==3.0.3
django-bootstrap-modal-forms==1.5.0
django-bootstrap3==12.1.0
django-bootstrap4==1.1.1
django-braces==1.14.0
django-db-file-storage==0.5.5
django-filter==2.0.0
django-formtools==2.2
django-heroku==0.3.1
django-payments==0.13.0
django-payments-razorpay==1.0
django-phone-field==1.8.1
django-sendsms==0.3.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drfaddons==0.1.3
ezenv==0.91
Faker==4.0.3
gunicorn==20.0.4
idna==2.9
misaka==2.1.1
Naked==0.1.31
paywix==1.3.0
phonenumbers==8.12.7
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
pycryptodome==3.4.3
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
razorpay==1.2.0
requests==2.23.0
shellescape==3.8.1
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.50.0
suds-jurko==0.6
text-unidecode==1.3
urllib3==1.25.9
whitenoise==5.2.0
wincertstore==0.2
xmltodict==0.12.0
